Add header to ChangeTeamDetailsPage and ResetPasswordPage
* Create a header component with just the retroquest logo, and the ability to pass in a name and children
* Use that compnent within the team header, and within the ChangeTeamDetailsPage, and ResetPasswordPage pages
