+++ ui/src/App/Team/TeamHeader/Settings/InfoTab/InfoTab.scss
