+++ ui/src/App/Team/TeamHeader/Settings/StylesTab/StylesTab.scss
