+++ ui/src/App/Team/TeamHeader/TeamHeader.scss
@@ -31,19 +31,6 @@
 		height: 64px;
 	}
 
-	.left-content {
-		display: flex;
-		flex: 1;
-		align-items: center;
-		overflow: hidden;
-
-		font-size: 1rem;
-
-		@include main.breakpoint-medium {
-			font-size: 1.5rem;
-		}
-	}
-
 	.center-content {
 		display: none;
 		flex: 1;
@@ -66,61 +53,6 @@
 		}
 	}
 
-	.logo-link {
-		font-size: 0;
-		border: 2px solid transparent;
-
-		&:hover,
-		&:focus {
-			border: 2px solid main.$blue;
-			border-radius: 7px;
-		}
-	}
-
-	.logo {
-		height: 24px;
-		border-radius: 6px;
-
-		@include main.breakpoint-medium {
-			height: 36px;
-		}
-	}
-
-	.horizontal-separator {
-		display: none;
-		width: 3px;
-		height: 24px;
-		margin: 0 24px;
-
-		background-color: main.$asphalt;
-		border-radius: 6px;
-		opacity: 0.1;
-
-		@include main.breakpoint-medium {
-			display: initial;
-		}
-	}
-
-	.team-name {
-		flex: 1;
-		padding: 0 6px;
-		overflow: hidden;
-
-		color: inherit;
-		font-weight: bold;
-		font-size: inherit;
-		white-space: nowrap;
-		text-align: center;
-		text-overflow: ellipsis;
-
-		@include main.breakpoint-medium {
-			flex: initial;
-			padding: initial;
-
-			text-align: initial;
-		}
-	}
-
 	.nav-link {
 		height: 28px;
 
@@ -133,22 +65,6 @@
 		&.selected {
 			color: main.$asphalt;
 			background-color: rgba(main.$asphalt, 0.2);
-		}
-	}
-
-	.last-saved-text {
-		display: none;
-		margin-right: 18px;
-
-		font-weight: 500;
-		font-size: 0.8rem;
-		white-space: nowrap;
-		text-align: right;
-
-		opacity: 0.5;
-
-		@include main.breakpoint-medium {
-			display: initial;
 		}
 	}
 
@@ -166,15 +82,6 @@
 	.header {
 		.right-content {
 			color: main.$gray-1;
-		}
-
-		.horizontal-separator {
-			background-color: rgba(main.$black, 0.22);
-			opacity: 1;
-		}
-
-		.team-name {
-			color: rgba(main.$gray-1, 0.75);
 		}
 
 		.nav-link {
