+++ ui/src/Common/Header/Header.spec.tsx
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { screen, within } from '@testing-library/react';
+
+import renderWithRecoilRoot from '../../Utils/renderWithRecoilRoot';
+
+import Header from './Header';
+
+describe('Header', () => {
+	beforeEach(() => {
+		renderWithRecoilRoot(<Header name="Test Name">children</Header>);
+	});
+
+	it('should render name', () => {
+		expect(screen.getByText('Test Name')).toBeDefined();
+	});
+
+	it('should render logo link', async () => {
+		const retroQuestLogoLink = screen.getByTestId('retroquestLogoLink');
+		expect(retroQuestLogoLink.getAttribute('href')).toBe('/');
+		expect(
+			within(retroQuestLogoLink).getByAltText('Retro Quest')
+		).toBeDefined();
+	});
+
+	it('should render children', () => {
+		expect(screen.getByText('children')).toBeDefined();
+	});
+});
